Fix resize for CAPI clusters
Clusters created by the magnum-capi-helm driver do not have a heat stack, so
resize currently fails as it depends on the heat stack for a list of worker
nodes. To fix this, we change to querying the nodegroup information and
use that instead.

We keep the old heat stack way for now, as that is the way to get a list
of worker nodes to select candidate(s) for scaling down. This is not
available in capi driver yet, so capi will return an empty list of
worker nodes.

Change-Id: I0d66fc02a7f2c608a4a0b09b98c343017e04ed41
(cherry picked from commit a5ff9ecb4fb84e95c266657b5aabdefe33010983)
(cherry picked from commit 19c1c153fbc1a2cf5808e77cfa5909b16456e023)